[FIX] #98 separate detail page components
